refactor: rearrange symbols
