+++ packages/core/src/types.ts
@@ -3,7 +3,7 @@ import { Either } from 'fp-ts/lib/Either';
 import { Reader } from 'fp-ts/lib/Reader';
 import { TaskEither } from 'fp-ts/lib/TaskEither';
 import { Logger } from 'pino';
-import { AuthErr } from './utils/security/handlers';
+import { AuthErr } from './utils/security/handlers/types';
 export type IPrismDiagnostic = Omit<IDiagnostic, 'range'>;
 
 // END
