+++ packages/core/src/utils/security/handlers/types.ts
@@ -1,0 +1,6 @@
+import { Dictionary } from '@stoplight/types';
+
+type Headers = Dictionary<string, string>;
+
+export type AuthErr = { name: string; message: string; status: number; headers: Headers };
+export type SecurityScheme = { type: string; name: string; in?: string; scheme?: string };
